Make `routing_key` accept multiple values
- Implement `RoutableProvider` for all the inference adapters
